+++ MoyaProvider+ReactiveSwift.swift
@@ -1,0 +1,3 @@
+import ReactiveSwift
+
+extension MoyaProvider: ReactiveExtensionsProvider { }