+++ Tests/MoyaProviderIntegrationTests.swift
@@ -253,9 +253,9 @@ class MoyaProviderIntegrationTests: QuickSpec {
             }
 
             describe("a reactive provider with SignalProducer") {
-                var provider: ReactiveSwiftMoyaProvider<GitHub>!
+                var provider: Reactive<MoyaProvider<GitHub>>!
                 beforeEach {
-                    provider = ReactiveSwiftMoyaProvider<GitHub>()
+                    provider = MoyaProvider<GitHub>().reactive
                 }
 
                 it("returns some data for zen request") {